requirements.txt
appdirs==1.4.4
attrs==23.1.0
cattrs==23.1.2
certifi==2023.7.22
charset-normalizer==3.2.0
contourpy==1.1.1
cycler==0.11.0
exceptiongroup==1.1.3
fonttools==4.42.1
idna==3.4
jplephem==2.19
kerykeion==4.2.4
kiwisolver==1.4.5
matplotlib==3.8.0
numpy==1.26.0
packaging==23.1
pandas==2.1.1
Pillow==10.0.1
pydantic==1.10.13
pyparsing==3.1.1
pyswisseph==2.10.3.2
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.31.0
requests-cache==0.9.8
sgp4==2.22
six==1.16.0
skyfield==1.46
terminaltables==3.1.10
typing_extensions==4.8.0
tzdata==2023.3
url-normalize==1.4.3
urllib3==2.0.5
